changes so far in moving directories
