+++ web/src/components/__tests__/Register.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, fireEvent, waitFor, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { FlowContext } from '../../FlowContext.js';
-import Register from '../base/Register';
+import Register from '../../features/auth/Register';
 import { useNavigate } from 'react-router-dom';
 import { Auth } from 'aws-amplify';
