+++ web/src/outdated/ui-components/AccommodationCreateForm.d.ts
