+++ web/src/outdated/ui-components/AccommodationUpdateForm.d.ts
