+++ web/src/outdated/ui-components/Property1Opened.d.ts
