+++ web/src/outdated/ui-components/Property1Opened.jsx
